Fixed dynamic complation errors. [#METR-22071]
