+++ dbms/include/DB/Interpreters/Aggregator.h
@@ -39,20 +39,20 @@ namespace ErrorCodes
 /** Разные структуры данных, которые могут использоваться для агрегации
   * Для эффективности, сами данные для агрегации кладутся в пул.
   * Владение данными (состояний агрегатных функций) и пулом
-  *	 захватывается позднее - в функции convertToBlocks, объектом ColumnAggregateFunction.
+  *  захватывается позднее - в функции convertToBlocks, объектом ColumnAggregateFunction.
   *
   * Большинство структур данных существует в двух вариантах: обычном и двухуровневом (TwoLevel).
   * Двухуровневая хэш-таблица работает чуть медленнее при маленьком количестве различных ключей,
-  *	 но при большом количестве различных ключей лучше масштабируется, так как позволяет
-  *	 распараллелить некоторые операции (слияние, пост-обработку) естественным образом.
+  *  но при большом количестве различных ключей лучше масштабируется, так как позволяет
+  * распараллелить некоторые операции (слияние, пост-обработку) естественным образом.
   *
   * Чтобы обеспечить эффективную работу в большом диапазоне условий,
-  *	 сначала используются одноуровневые хэш-таблицы,
-  *	 а при достижении количеством различных ключей достаточно большого размера,
-  *	 они конвертируются в двухуровневые.
+  *  сначала используются одноуровневые хэш-таблицы,
+  *  а при достижении количеством различных ключей достаточно большого размера,
+  *  они конвертируются в двухуровневые.
   *
   * PS. Существует много различных подходов к эффективной реализации параллельной и распределённой агрегации,
-  *	 лучшим образом подходящих для разных случаев, и этот подход - всего лишь один из них, выбранный по совокупности причин.
+  *  лучшим образом подходящих для разных случаев, и этот подход - всего лишь один из них, выбранный по совокупности причин.
   */
 using AggregatedDataWithoutKey = AggregateDataPtr;
 
@@ -384,7 +384,7 @@ struct AggregationMethodConcat
 		{
 			/** Исправление, если все ключи - пустые массивы. Для них в хэш-таблицу записывается StringRef нулевой длины, но с ненулевым указателем.
 			  * Но при вставке в хэш-таблицу, такой StringRef оказывается равен другому ключу нулевой длины,
-			  *	 у которого указатель на данные может быть любым мусором и использовать его нельзя.
+			  * у которого указатель на данные может быть любым мусором и использовать его нельзя.
 			  */
 			for (size_t i = 0; i < keys_size; ++i)
 				key_columns[i]->insertDefault();
@@ -525,14 +525,14 @@ struct AggregatedDataVariants : private boost::noncopyable
 	  * - при агрегации, состояния создаются в пуле с помощью функции IAggregateFunction::create (внутри - placement new произвольной структуры);
 	  * - они должны быть затем уничтожены с помощью IAggregateFunction::destroy (внутри - вызов деструктора произвольной структуры);
 	  * - если агрегация завершена, то, в функции Aggregator::convertToBlocks, указатели на состояния агрегатных функций
-	  *	  записываются в ColumnAggregateFunction; ColumnAggregateFunction "захватывает владение" ими, то есть - вызывает destroy в своём деструкторе.
+	  *   записываются в ColumnAggregateFunction; ColumnAggregateFunction "захватывает владение" ими, то есть - вызывает destroy в своём деструкторе.
 	  * - если при агрегации, до вызова Aggregator::convertToBlocks вылетело исключение,
-	  *	  то состояния агрегатных функций всё-равно должны быть уничтожены,
-	  *	  иначе для сложных состояний (наприемер, AggregateFunctionUniq), будут утечки памяти;
+	  *   то состояния агрегатных функций всё-равно должны быть уничтожены,
+	  *   иначе для сложных состояний (наприемер, AggregateFunctionUniq), будут утечки памяти;
 	  * - чтобы, в этом случае, уничтожить состояния, в деструкторе вызывается метод Aggregator::destroyAggregateStates,
-	  *	  но только если переменная aggregator (см. ниже) не nullptr;
+	  *   но только если переменная aggregator (см. ниже) не nullptr;
 	  * - то есть, пока вы не передали владение состояниями агрегатных функций в ColumnAggregateFunction, установите переменную aggregator,
-	  *	  чтобы при возникновении исключения, состояния были корректно уничтожены.
+	  *   чтобы при возникновении исключения, состояния были корректно уничтожены.
 	  *
 	  * PS. Это можно исправить, сделав пул, который знает о том, какие состояния агрегатных функций и в каком порядке в него уложены, и умеет сам их уничтожать.
 	  * Но это вряд ли можно просто сделать, так как в этот же пул планируется класть строки переменной длины.
@@ -763,7 +763,7 @@ using ManyAggregatedDataVariants = std::vector<AggregatedDataVariantsPtr>;
   * В случае наличия group_by_overflow_mode = 'any', данные агрегируются как обычно, кроме ключей, не поместившихся в max_rows_to_group_by.
   * Для этих ключей, данные агрегируются в одну дополнительную строку - далее см. под названиями overflow_row, overflows...
   * Позже, состояния агрегатных функций для всех строк (прошедших через HAVING) мерджатся в одну,
-  *	 а также к ним прибавляется или не прибавляется (в зависимости от настройки totals_mode) также overflow_row - это и будет TOTALS.
+  *  а также к ним прибавляется или не прибавляется (в зависимости от настройки totals_mode) также overflow_row - это и будет TOTALS.
   */
 
 
@@ -792,7 +792,7 @@ public:
 
 		/// Настройки двухуровневой агрегации (используется для большого количества ключей).
 		/** При каком количестве ключей или размере состояния агрегации в байтах,
-		  *	 начинает использоваться двухуровневая агрегация. Достаточно срабатывания хотя бы одного из порогов.
+		  *  начинает использоваться двухуровневая агрегация. Достаточно срабатывания хотя бы одного из порогов.
 		  * 0 - соответствующий порог не задан.
 		  */
 		const size_t group_by_two_level_threshold;
@@ -850,7 +850,7 @@ public:
 	  * Если overflow_row = true, то агрегаты для строк, не попавших в max_rows_to_group_by, кладутся в первый блок.
 	  *
 	  * Если final = false, то в качестве столбцов-агрегатов создаются ColumnAggregateFunction с состоянием вычислений,
-	  *	 которые могут быть затем объединены с другими состояниями (для распределённой обработки запроса).
+	  *  которые могут быть затем объединены с другими состояниями (для распределённой обработки запроса).
 	  * Если final = true, то в качестве столбцов-агрегатов создаются столбцы с готовыми значениями.
 	  */
 	BlocksList convertToBlocks(AggregatedDataVariants & data_variants, bool final, size_t max_threads) const;
@@ -917,8 +917,8 @@ protected:
 	  * Внутренний цикл (для случая without_key) получается почти в два раза компактнее; прирост производительности около 30%.
 	  *
 	  * 2. Вызов по указателю на функцию лучше, чем виртуальный вызов, потому что в случае виртуального вызова,
-	  *	 GCC 5.1.2 генерирует код, который на каждой итерации цикла заново грузит из памяти в регистр адрес функции
-	  *	 (значение по смещению в таблице виртуальных функций).
+	  *  GCC 5.1.2 генерирует код, который на каждой итерации цикла заново грузит из памяти в регистр адрес функции
+	  *  (значение по смещению в таблице виртуальных функций).
 	  */
 	struct AggregateFunctionInstruction
 	{
@@ -947,9 +947,9 @@ protected:
 
 	/** Динамически скомпилированная библиотека для агрегации, если есть.
 	  * Смысл динамической компиляции в том, чтобы специализировать код
-	  *	 под конкретный список агрегатных функций.
+	  *  под конкретный список агрегатных функций.
 	  * Это позволяет развернуть цикл по созданию и обновлению состояний агрегатных функций,
-	  *	 а также использовать вместо виртуальных вызовов inline-код.
+	  *  а также использовать вместо виртуальных вызовов inline-код.
 	  */
 	struct CompiledData
 	{
@@ -977,7 +977,7 @@ protected:
 	void initialize(const Block & block);
 
 	/** Установить блок - пример результата,
-	  *	 только если он ещё не был установлен.
+	  * только если он ещё не был установлен.
 	  */
 	void setSampleBlock(const Block & block);
 
@@ -1065,7 +1065,8 @@ public:
 	void executeSpecializedWithoutKey(
 		AggregatedDataWithoutKey & res,
 		size_t rows,
-		AggregateColumns & aggregate_columns) const;
+		AggregateColumns & aggregate_columns,
+		Arena * arena) const;
 
 protected:
 	/// Слить данные из хэш-таблицы src в dst.
